refactor: consolidate Card mutation API
