+++ src/entities/card/api/mutations.ts
@@ -1,4 +1,4 @@
-import type { CardCreateInput, CardId, DeleteCardInput, EditCardInput } from "../model/types";
+import type { CardCreateInput, CardEdit, CardId, DeleteCardInput, EditCardInput } from "../model/types";
 
 import { findDeckById } from "@/entities/deck/@x/card";
 import { createLocalCard, deleteLocalCard, editLocalCard, findCardById } from "../model/store";
@@ -7,6 +7,18 @@ import {
   deleteCard as deleteRemoteCard,
   editCard as editRemoteCard,
 } from "./firestore";
+
+export type CardMutation = { kind: "create"; card: CardCreateInput } | { kind: "edit"; card: CardEdit };
+
+export class CardBulkMutationError extends Error {
+  constructor(
+    public readonly failedIds: CardId[],
+    total: number,
+    options?: ErrorOptions
+  ) {
+    super(`${failedIds.length} of ${total} Card writes failed`, options);
+  }
+}
 
 const isLocalDeck = (deckId: string): boolean => findDeckById(deckId)?.localMode === true;
 
@@ -22,7 +34,7 @@ const requireLocalMode = (deckId: string): boolean => {
   return deck.localMode;
 };
 
-export const createCard = async (uid: string, card: CardCreateInput): Promise<void> => {
+const createCard = async (uid: string, card: CardCreateInput): Promise<void> => {
   if (isLocalDeck(card.deckId)) {
     createLocalCard(card);
     return;
@@ -37,6 +49,19 @@ export const editCard = async (uid: string, card: EditCardInput["card"]): Promis
     return;
   }
   await editRemoteCard(uid, card);
+};
+
+export const mutateCards = async (uid: string, mutations: CardMutation[]): Promise<void> => {
+  const results = await Promise.allSettled(
+    mutations.map((mutation) =>
+      mutation.kind === "create" ? createCard(uid, mutation.card) : editCard(uid, mutation.card)
+    )
+  );
+  const failedIds = results.flatMap((result, index) => {
+    const mutation = mutations[index];
+    return result.status === "rejected" && mutation != null ? [mutation.card.id] : [];
+  });
+  if (failedIds.length > 0) throw new CardBulkMutationError(failedIds, mutations.length);
 };
 
 export const deleteCard = async (uid: string, card: DeleteCardInput["card"]): Promise<void> => {
