+++ src/entities/card/index.ts
@@ -1,6 +1,6 @@
 export { fetchCards, subscribeCards } from "./api/firestore";
-export { deleteCard, editCard } from "./api/mutations";
-export { CardBulkMutationError, mutateCards } from "./api/mutateCards";
+export { CardBulkMutationError, deleteCard, editCard, mutateCards } from "./api/mutations";
+export type { CardMutation } from "./api/mutations";
 export { generateCardId } from "./model/id";
 export { useCard, useCards, useCardsByDeckId } from "./model/hooks";
 export { clearRemoteCards } from "./model/store";
@@ -15,7 +15,6 @@ export type {
   Card,
   CardEdit,
   CardId,
-  CardMutation,
   CardRaw,
 } from "./model/types";
 export { filterCardsByDeckId, mustFindCardById } from "./model/rules";
