+++ src/entities/card/model/types.ts
@@ -14,7 +14,6 @@ export type CardCreate = z.infer<typeof cardCreateSchema>;
 export type CardCreateInput = z.input<typeof cardCreateSchema>;
 export type CardId = z.infer<typeof cardIdSchema>;
 export type CardEdit = z.infer<typeof cardEditSchema>;
-export type CardMutation = { kind: "create"; card: CardCreateInput } | { kind: "edit"; card: CardEdit };
 export type CardRaw = Pick<Card, "frontText" | "backText" | "uniqueKey" | "tags">;
 export type EditCardInput = z.infer<typeof editCardSchema>;
 export type DeleteCardInput = z.infer<typeof deleteCardSchema>;
